requirements.txt
pydantic>=2.5.0
pydantic-ai>=0.1.0
python-dotenv>=1.0.0
logfire>=0.8.0
anthropic>=0.8.0
openai>=1.0.0
mcp-server-fetch>=0.1.0
rich>=13.0.0
aiohttp>=3.9.0

# Testing dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0 